requirements.txt
asgiref==3.5.0
distlib==0.3.4
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
filelock==3.6.0
gunicorn==20.1.0
platformdirs==2.5.1
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
virtualenv==20.13.3
virtualenv-clone==0.5.7
whitenoise==6.0.0
